refactor: rename the relationships client producer
Since we are going to introduce another producer, it makes sense to have
the producer for the relationships in its own package.

Since I was at it, I improved some log messages to make it clear when we
are generating clients for the relationships API, and included an
explanation at the top of the class clarifying why we implemented our
own producers.

RHCLOUD-34485