RHCLOUD-32983 | Add support for the Kessel authorization in the back end (#2788)
* feature: support the Kessel authentication in the back end endpoints

The goal is to be able to use Kessel instead of RBAC for when we are
trying to figure out the user's authorization for an asset or endpoint
they are trying to access.

RHCLOUD-32983

* refactor: remove unused security contexts

RHCLOUD-32983

* refactor: delay validation of the payloads

We are interested in checking for authorization first, and then we can
go and validate if the body has been properly sent or not.

RHCLOUD-32983

* refactor: create an exception mapper for the constraint violations

For some reason the constraint violations that are thrown from methods
are not picked up by the default exception mapper. Therefore, we need to
create a specific one for these kind of exceptions.

RHCLOUD-32983

* refactor: permissions for the integrations

After talking to the AuthZ team, we have figured out which permissions
we are going to use for the back end. And for now, we are going to just
put the integrations behind the new authorization system.

RHCLOUD-32983

* fix: typo

RHCLOUD-32983

* refactor: include the authorized IDs in the query when there are none

The query was not including the "authorized_ids IN" condition if the
given authorized IDs list was empty, which was making all the endpoints
of that organization to be returned. A regression test has been added to
avoid that.

RHCLOUD-32983

* refactor: make methods private where possible

Some of them need to be at least "protected" so that the validations and
the roles can work.

RHCLOUD-32983

* refactor: modify the permission names

Kessel is going to use the "subscribe" verb for the permissions, so we
needed to update them here.

RHCLOUD-32983

* feature: log when no integration IDs have been returned

When the subject is not authorized to fetch any of the workspace's
integrations or when there are none to return, we simply respond with an
empty integrations page for the user.

Since Kessel is a fairly new integration, we think that we need to log
this behavior so that we can debug and understand it when it happens at
these initial stages.

RHCLOUD-32983