feature: migration resource for Errata subscriptions
We need to migrate all the customer's subscriptions from the Errata's
legacy system to our database, and for that, we are going to use a
resource that allows us to upload either a CSV or a JSON file with the
subscription contents.

We will subscribe every given user to all the Errata event types.

RHCLOUD-34459