feature: return "not found" responses for unauthorized callers
In order to return consistent responses when an unauthorized principal
calls our APIs, we have decided as a team that we will return "not
found" responses.

The reasoning is that many existing services will not let you know
whether the resource exists or not if you are not authorized, so we
thought that following this pattern was a good idea.

RHCLOUD-35763