refactor: only fetch regular endpoints
We want to leave the system endpoints outside of the migration because
they do not belong to any organization in particular and are only
managed by the Notifications team. Since those endpoints do not have any
account or organization identifiers, they will not be part of any
workspace anyway.

RHCLOUD-35453