RHCLOUD-37430 | fix: integrations are not being returned sometimes (#3256)
* fix: integrations are not being returned sometimes

We were getting reports of integrations not showing up in stage when
Kessel Relations was enabled. More specifically, the issue was that
after trying to filter integrations by name, some of them did simply not
show up in the results even though they were recently created.

After investigating the issue, it turned out that the problem was that we
were getting an incomplete set of authorized integration IDs from Kessel,
which in turn made us filter the database query results incorrectly, which
again, in turn, made us return an inaccurate list of integrations for an
organization.

The QE organization which reported the issue had around 2,500
integrations created, and when a request was made to look up for a very
specific integration with name "X", Notifications first had to ask
Kessel for the principal's authorized integrations. Since Kessel only
returned a batch of a 1,000 we would use just those to apply a
"integration ID in" filter in the query, so if the specified name in the
request was in that batch, the user was lucky and the integration was
returned.

These changes introduce a mechanism that keeps firing requests to Kessel
whenever there is a continuation token available. That makes sure that
we are fetching all the integrations that the user has access to, so
that we don't miss any when querying our database.

RHCLOUD-37430

* feature: implement post-filtering methods for Kessel

Even though we are not going to use them now, I already developed the
code for the "post-filtering" tests on a different PR. We might consider
this approach in the future, so it is worth having the code already
developed.

RHCLOUD-37430

* refactor: remove debugging log traces

We are no longer going to use those log traces, so we can safely remove
them from the code.

RHCLOUD-37430

* feature: allow configuring the limit when looking up resources

Aligns with Kessel's limit and allows us to modify it in case they also
change it.

RHCLOUD-37430