feature: bypass the maximum behavior group creation limit with Unleash
There are very specific organizations that, for reasons that make sense,
might want to have more than the currently established limit of
behavior groups.

For that purpose, we are simply adding an Unleash toggle that will be
enabled by default for everybody, and that will allow us to selectively
disable it for certain organizations.

RHCLOUD-37680