feature: authorization annotation for improved maintainability
The implemented annotation allows specifying the legacy RBAC role that
will be used when RBAC is enabled, and the Kessel permissions for when
that one is being used as the authorization back end.

The goal is to reduce the boilerplate and repetitive code that we have
in the code base, and make it easier for the developers to add
authorization in a similar way that it is done with the "RolesAllowed"
annotation.