RHCLOUD-38438 | fix: Turnpike requests unauthorized when Kessel is enabled (#3314)
* fix: Turnpike requests unauthorized when Kessel is enabled

We realized that when enabling Kessel Relations in stage or production,
and when sending Turnpike authenticated requests to the internal
endpoints, these would simply return "Forbidden" responses to the
caller.

The problem was that for Turnpike requests, for some reason, the Kessel
Relations back end was being signaled as "enabled", which ended up
building a security identity without the required internal roles for the
internal endpoints to work.

Since Kessel only supports "service account" and "user" principals,
adding a check for those before proceeding with building a security
identity tailored for Kessel solves the issue.

Now the Turnpike authenticated requests will be identified as such when
building the security identity, and the proper roles will be added to
it.

RHCLOUD-38438

* refactor: expand code for improved readability

RHCLOUD-38438

* refactor: add the log statement back for printing the identity header

RHCLOUD-38438